+++ src/Ext/WeaponType/Body.h
@@ -45,6 +45,7 @@ public:
 		Valueable<AreaFireTarget> AreaFire_Target;
 		Valueable<WeaponTypeClass*> FeedbackWeapon;
 		Valueable<bool> Laser_IsSingleColor;
+		Valueable<PositionFollow> LaserPositionUpdate;
 		Nullable<int> LaserZAdjust;
 		Nullable<int> EBoltZAdjust;
 		Nullable<bool> EBoltZAdjust_ClampInitialDepthForBuilding;
@@ -134,6 +135,7 @@ public:
 			, AreaFire_Target { AreaFireTarget::Base }
 			, FeedbackWeapon {}
 			, Laser_IsSingleColor { false }
+			, LaserPositionUpdate { PositionFollow::None }
 			, LaserZAdjust {}
 			, EBoltZAdjust {}
 			, EBoltZAdjust_ClampInitialDepthForBuilding {}
@@ -231,4 +233,7 @@ public:
 	static int GetRangeWithModifiers(WeaponTypeClass* pThis, TechnoClass* pFirer);
 	static int GetRangeWithModifiers(WeaponTypeClass* pThis, TechnoClass* pFirer, int range);
 	static int GetTechnoKeepRange(WeaponTypeClass* pThis, TechnoClass* pFirer, bool isMinimum);
+
+	// Misc/Hooks.LaserDraw.cpp
+	static void LaserTrackingPointerExpired(void* ptr, bool removed);
 };